updated to allow adding strains to existing folder

+++ PyParanoid.py
@@ -3,7 +3,7 @@
 #UBC Microbiology - Haney Lab
 
 import argparse, os, errno, shutil, subprocess
-import itertools
+import itertools, sys
 from Bio import SeqIO
 
 def parse_args():
@@ -15,6 +15,7 @@ Note that default diamond uses 6 threads when requesting resources.
 	parser.add_argument('genomedb',type=str,help='relative path to genomedb for raw data')
 	parser.add_argument('strainlist',type=str,help='text file, one strain per line. names should be "species" field from genome_metadata table')
 	parser.add_argument('outdir',type=str,help='folder to store output')
+	parser.add_argument('--add',action='store_true',help='use if you intend to add to an existing PyParanoid output folder')
 	return parser.parse_args()
 
 def setupdir(outdir,strains,genomedb):
@@ -35,7 +36,8 @@ def setupdir(outdir,strains,genomedb):
 			shutil.copy(os.path.join(genomedb,"pep",s+".pep.fa"),os.path.join(outdir,"faa",s+".faa"))
 		except IOError as exc:
 			if exc.errno == 2:
-				print s, 'not found in database...'
+				print s, 'not found in database...check your strainlist.'
+				sys.exit()
 	return
 
 def make_diamond_databases(strains,outdir):
@@ -85,6 +87,42 @@ def run_diamond(strains,outdir):
 			pass
 	return
 
+def run_diamond_on_old_strains(strains,outdir,old_strains):
+
+	print "Running reflexive diamond on all", len(strains), "strains..."
+	count = len(strains)
+	for s in old_strains:
+		cmds = "diamond blastp --query {}/faa/{}.faa -d {}/dmnd/{}.dmnd -o {}/m8/{}.{}.m8 -f tab --min-score 50 --quiet --threads 6".format(outdir,s,outdir,s,outdir,s,s)
+		proc = subprocess.Popen(cmds.split())
+		proc.wait()
+		count -= 1
+		if count == 0:
+			print "\tDone!"
+		elif (count % 10 == 0):
+			print "\t"+str(count), "remaining..."
+		else:
+			pass
+
+	count = 2 * len(strains) * len(old_strains)
+	print "Running pairwise diamond against all old strains:", count, "pairwise permutations ..."
+	for s in strains:
+		for t in old_strains:
+			cmds = "diamond blastp --query {}/faa/{}.faa -d {}/dmnd/{}.dmnd -o {}/m8/{}.{}.m8 -f tab --min-score 50 --quiet --threads 6".format(outdir,s,outdir,t,outdir,s,t)
+			proc = subprocess.Popen(cmds.split())
+			proc.wait()
+			cmds = "diamond blastp --query {}/faa/{}.faa -d {}/dmnd/{}.dmnd -o {}/m8/{}.{}.m8 -f tab --min-score 50 --quiet --threads 6".format(outdir,t,outdir,s,outdir,t,s)
+			proc = subprocess.Popen(cmds.split())
+			proc.wait()
+			count -= 2
+			if count == 0:
+				print "\tDone!"
+			elif count % 10 == 0:
+				print "\t"+str(count), "remaining..."
+			else:
+				pass
+	return
+
+
 def parse_diamond(genes,outdir):
 	print "Parsing diamond results..."
 	files = os.listdir(os.path.join(outdir,"m8"))
@@ -122,13 +160,23 @@ def get_genes(strains,outdir):
 	return genes
 
 def run_inparanoid(strains,outdir):
-	pairs = list(itertools.permutations(strains,2))
+
+	pairs = list(itertools.combinations(strains,2))
 	count = len(pairs)
-	print "Running InParanoid to identify orthologs for",count,"permutations..."
+	print "Running inparanoid on", count, "pairs of strains..."
 	for p in pairs:
-		cmds = "perl inparanoid2.pl {} {} {}".format(p[0],p[1],outdir,outdir)
-		proc = subprocess.Popen(cmds.split())
-		proc.wait()
+		a = "{}.{}.txt".format(p[0],p[1])
+		b = "{}.{}.txt".format(p[1],p[0])
+		if p[0] == p[1]:
+			pass
+		elif a in os.listdir(os.path.join(outdir,"paranoid_output")):
+			pass
+		elif b in os.listdir(os.path.join(outdir,"paranoid_output")):
+			pass
+		else:
+			cmds = "perl inparanoid2.pl {} {} {}".format(p[0],p[1],outdir)
+			proc = subprocess.Popen(cmds.split())
+			proc.wait()
 		count -= 1
 		if count == 0:
 			print "\tDone!"
@@ -142,7 +190,25 @@ def clean_up(outdir):
 	for f in os.listdir(os.path.join(outdir,"m8")):
 		os.remove(os.path.join(outdir,"m8",f))
 	for f in os.listdir(os.path.join(outdir,"out")):
-		os.remove(os.path.join(outdir,"out",f))
+		vals = f.split(".")
+		if vals[0] == vals[1]:
+			pass
+		else:
+			os.remove(os.path.join(outdir,"out",f))
+	return
+
+def append_strains(outdir,strainlist):
+	o = open(os.path.join(outdir,"strainlist.txt"),'a')
+	for line in open(strainlist,'r'):
+		o.write(line)
+	return
+
+def check_unique(old_strains,strains):
+	for s in strains:
+		if s in old_strains:
+			print "New strain already in database. Check strainlist and try again. Exiting..."
+		else:
+			pass
 	return
 
 def main():
@@ -151,14 +217,30 @@ def main():
 	strains = [x.rstrip() for x in open(os.path.abspath(args.strainlist),'r')]
 	outdir = os.path.abspath(args.outdir)
 
-	setupdir(outdir,strains,genomedb)
-	shutil.copy(os.path.abspath(args.strainlist),os.path.join(outdir,"strainlist.txt"))
-	make_diamond_databases(strains,outdir)
-	run_diamond(strains,outdir)
-	genes = get_genes(strains,outdir)
-	parse_diamond(genes,outdir)
-	run_inparanoid(strains, outdir)
-	clean_up(outdir)
+	if args.add:
+		print "adding"
+		setupdir(outdir,strains,genomedb)
+		old_strains = [x.rstrip() for x in open(os.path.join(outdir,"strainlist.txt"),'r')]
+		check_unique(old_strains,strains)
+		append_strains(outdir,os.path.abspath(args.strainlist))
+		make_diamond_databases(strains,outdir)
+		run_diamond(strains,outdir)
+		run_diamond_on_old_strains(strains,outdir,old_strains)
+		[strains.append(s) for s in old_strains]
+		genes = get_genes(strains,outdir)
+		parse_diamond(genes,outdir)
+		run_inparanoid(strains, outdir)
+		clean_up(outdir)
+	else:
+		print "not adding"
+		setupdir(outdir,strains,genomedb)
+		shutil.copy(os.path.abspath(args.strainlist),os.path.join(outdir,"strainlist.txt"))
+		make_diamond_databases(strains,outdir)
+		run_diamond(strains,outdir)
+		genes = get_genes(strains,outdir)
+		parse_diamond(genes,outdir)
+		run_inparanoid(strains, outdir)
+		clean_up(outdir)
 
 if __name__ == '__main__':
 	main()
